Converge __format() toward Glibc extensions.
Add user-specified field width.

Add '#' case-change flag (applies only to certain specs, but trumps '^'
if it does apply).

Unrecognized specifications are left as-is, except for the actions of
flags. So '%.' remains '%.', rather than becoming '.' as the BSD
functionality would have it. And '%Ez' remains '%Ez' rather than being
interpreted the same as '%z'. But '%^Ez' becomes '%^EZ', which looks
like a bug to me, but that's what Glibc strftime() does.

Add '%v' (OK, that's a BSD extension, but I could not resist).