Add rudimentary argument validation.
The incentive is that I might want to bolt on a locale system to
DateTime::Fiction::JRRTolkien::Shire, and that might (or might not)
involve adding arguments. So I want to flag extra arguments now, to
reserve them for myself later.

I do not want to use a type system because I want this module to be very
light weight, and to work back as far as Date::Tolkien::Shire does. Now,
I set that to 5.6.2 when I adopted it (before, it was unspecified), but
that seems about right given the state of the code.