Modify methods expected by __format in date obj
Date::Tolkien::Shire and DateTime::Fiction::JRRTolkien::Shire have
equivocated on the semantics of month(), weekday(), and holiday(); the
former returns a name, the latter a number (or nothing for weekday(),
which it does not use).

With this commit __format expects the date object to support methods
month_number(), weekday_number(), and holiday_number() respectively.
This works in both the targeted packages since none of these is in use.

I have also made year() and day() into year_number() and day_number()
respectively, just for consistency's sake, even though there was no
equivocation on them.